refactor(auth): move auth-user.entity to auth domain directory
Update all imports to reference the new location of AuthUserEntity
Add new auth-user.entity implementation with tests
