+++ src/domain/auth/entities/__test__/auth-user.entity.test.ts
@@ -1,5 +1,5 @@
-import { AuthUserEntity } from "../auth-user.entity";
-import { UserEntity } from "../user.entity";
+import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
+import { UserEntity } from "@/domain/entities/user.entity";
 import { clearAllMocks } from "@/config/tests/test-utils";
 
 describe("AuthUserEntity", () => {
