+++ src/domain/auth/entities/auth-user.entity.ts
