+++ src/domain/auth/repositories/auth.repository.ts
@@ -3,7 +3,7 @@ import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { RegisterDto } from "@/domain/auth/dtos/register.dto";
 import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-password-email.dto";
 import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
-import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
+import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
 import { UserEntity } from "@/domain/entities/user.entity";
 
 export abstract class AuthRepository {
