+++ src/domain/auth/use-cases/__test__/register-user.test.ts
@@ -1,6 +1,6 @@
 import { RegisterUser } from "../register-user";
 import { RegisterDto } from "../../dtos/register.dto";
-import { AuthUserEntity } from "../../../entities/auth-user.entity";
+import { AuthUserEntity } from "../../entities/auth-user.entity";
 import { UserEntity } from "../../../entities/user.entity";
 import { MockAuthRepository } from "../../../../config/__tests__/mocks/auth-repository.mock";
