+++ src/domain/auth/use-cases/login-user.ts
@@ -1,5 +1,5 @@
 import { LoginDto } from "@/domain/auth/dtos/login.dto";
-import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
+import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
 import { AuthRepository } from "@/domain/auth/repositories/auth.repository";
 
 interface ILoginUserUseCase {
