+++ src/domain/user/use-cases/update-user-password.ts
@@ -1,5 +1,5 @@
 import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
-import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
+import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
 import { UserEntity } from "@/domain/entities/user.entity";
 import { AuthRepository } from "@/domain/auth/repositories/auth.repository";
