+++ src/infrastructure/datasources/auth.datasource.ts
@@ -1,5 +1,5 @@
 import { RegisterDto } from "@/domain/auth/dtos/register.dto";
-import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
+import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
 import { UserEntity } from "@/domain/entities/user.entity";
 import { AuthRepository } from "@/domain/auth/repositories/auth.repository";
 import { AuthClient } from "../config/auth.client";
